refactor: introduce step runner
